golang dependency injection and traefik
