+++ traefik/traefik.toml
@@ -1,0 +1,32 @@
+# traefik.toml
+
+[log]
+  level = "INFO"
+
+[entryPoints]
+  [entryPoints.http]
+    address = ":80"
+
+[api]
+  dashboard = true
+  insecure = true
+
+[providers]
+  [providers.file]
+    filename = "/etc/traefik/traefik.toml"
+
+[http]
+  [http.services]
+    [http.services.my-service.loadBalancer]
+      [[http.services.my-service.loadBalancer.servers]]
+        url = "http://app1:8080"
+      [[http.services.my-service.loadBalancer.servers]]
+        url = "http://app2:8080"
+      [[http.services.my-service.loadBalancer.servers]]
+        url = "http://app3:8080"
+
+  [http.routers]
+    [http.routers.my-router]
+      rule = "Host(`artificer.localhost.local`)"
+      service = "my-service"
+      entryPoints = ["http"]
